Crate tor_hsclient

tor-hsclient

Core implementation for onion services client.

This crate creates circuits to onion circuits, retains them for reuse, and hands them out as appropriate. It is also responsibile for maintaining all relevant state about hidden services, and their descriptors and introduction points.

Structs

DescriptorError

Error that occurred attempting to download a descriptor

HsClientConnector

An object that negotiates connections with onion services

HsClientSecretKeySpecifier

An identifier for a particular instance of an HS client key.

HsClientSecretKeys

Keys (if any) to use when connecting to a specific onion service.

HsClientSecretKeysBuilder

Builder for HsClientSecretKeys

HsClientSpecifier

An HS client identifier.

Enums

ConnError

Error that occurred attempting to reach a hidden service

DescriptorErrorDetail

Error that occurred attempting to download a descriptor

FailedAttemptError

Error that occurred making one attempt to connect to a hidden service using an IP and RP

HsClientKeyRole

The role of an HS client key.

InvalidTarget

We were given unusable information about an introduction point or rendezvous point.

StartupError

Error that occurred attempting to start up a hidden service client connector

Traits

HsClientConnectorConfig

Configuration for an HS client connector